Fix bugs to allow for automatic prompt grouping
Using the gemini API, the code can now automatically split up neutral
and opinionated residuals from the model.

The bug fix just changed the system prompt to not be RED/BLUE/Red/Blue,
but instead to be neutral/opinionated.

Also, a boolean switch for chat/non-chat llms was introduced, as well.