Fix issues with vector combination
When calculating vectors, the old behavior was for neutral vectors to be
the opposite of opinion vectors.

However, when combining vectors, the steering vector became
steer_vec = opinion_vec + sensible_vec
So, when calculating the neutral vector, in the past it was
neutral_vec = -steer_vec = -opinion_vec - sensible_vec

As can be seen, this leads to the neutral vector being the opposite of
sensible, a behavior which we do not desire.

Now the neutral vec = -opinion_vec + sensible_vec.

Also, the ratio of opinion_vec to sensible_vec was changed from 1:1 to
4:1 to make sure the opinionated behavior is most emphasized here.

Also, some additional testing and logging was performed.